Added "Custom Limelight" File
The file is just a copy of the limelight documentation, I plan on commenting the uses of each function in simple terms and removing anything that is not useful/redundant. Probably will be done in chunks as there is no way I can do this is one sitting. Feel free to add comments on functions but do not delete anything yet.